Undoing optin response renaming
