+++ spec/components/schemas/content/optin_response.ts
